consistently getter for size information in all `img`-classes
every image class has a the two getters:
std::string info() const;
size_t totalNbElements() const;

+++ modules/src/img/projectiondata.h
@@ -28,8 +28,9 @@ public:
 
         bool operator==(const Dimensions& other) const;
         bool operator!=(const Dimensions& other) const;
+
         std::string info() const;
-        uint totalNbElements() const;
+        size_t totalNbElements() const;
     };
 
     explicit ProjectionData(const SingleViewData::Dimensions& viewDimensions);
